pip-requirements.txt
aniso8601==3.0.0
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
decorator==4.3.0
Django==2.0.5
django-channels==0.7.0
Flask==1.0.2
hearthstone==4.3.5
hearthstone-data==24118.1
idna==2.6
ipython==6.3.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.0.7
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
pyinotify==0.9.6
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
